Moved Injector into main MoreDetails folder
